requirements.txt
streamlit==1.32.0
openai>=1.12.0
duckduckgo-search==4.4.3
python-dotenv==1.0.1
plotly==5.18.0
networkx==3.2.1 